+++ fwdpy11/headers/fwdpy11/discrete_demography/DiscreteDemography.hpp
@@ -62,8 +62,7 @@ namespace fwdpy11
 
         class demographic_model_state;
 
-        using demographic_model_state_pointer
-            = std::unique_ptr<demographic_model_state>;
+        using demographic_model_state_pointer = std::unique_ptr<demographic_model_state>;
 
         class DiscreteDemography
         {
@@ -105,12 +104,10 @@ namespace fwdpy11
             {
                 for (std::size_t i = 1; i < v.size(); ++i)
                     {
-                        if (v[i].when == v[i - 1].when
-                            && v[i].deme == v[i - 1].deme)
+                        if (v[i].when == v[i - 1].when && v[i].deme == v[i - 1].deme)
                             {
                                 std::ostringstream o;
-                                o << "DiscreteDemography: multiple "
-                                  << event_name(v[i])
+                                o << "DiscreteDemography: multiple " << event_name(v[i])
                                   << " events the same deme in the same "
                                      "generation";
                                 throw std::invalid_argument(o.str());
@@ -125,11 +122,9 @@ namespace fwdpy11
             {
                 for (std::size_t i = 1; i < v.size(); ++i)
                     {
-                        if (v[i].when == v[i - 1].when
-                            && v[i].source == v[i - 1].source
+                        if (v[i].when == v[i - 1].when && v[i].source == v[i - 1].source
                             && v[i].destination == v[i - 1].destination
-                            && v[i].move_individuals
-                                   == v[i - 1].move_individuals)
+                            && v[i].move_individuals == v[i - 1].move_individuals)
                             {
                                 throw std::invalid_argument(
                                     "DiscreteDemography: multiple "
@@ -147,8 +142,7 @@ namespace fwdpy11
                                 auto j = i + 1;
                                 double sum = i->fraction;
                                 for (; j < end(v) && j->source == i->source
-                                       && i->when == j->when
-                                       && j->move_individuals;
+                                       && i->when == j->when && j->move_individuals;
                                      ++j)
                                     {
                                         if (i->source == j->source)
@@ -159,10 +153,9 @@ namespace fwdpy11
                                 if (sum > 1.0)
                                     {
                                         std::ostringstream o;
-                                        o << "DiscreteDemography: at time "
-                                          << i->when << ", attempting to move "
-                                          << sum * 100.0 << "% of deme "
-                                          << i->source << " is invalid";
+                                        o << "DiscreteDemography: at time " << i->when
+                                          << ", attempting to move " << sum * 100.0
+                                          << "% of deme " << i->source << " is invalid";
                                         throw std::invalid_argument(o.str());
                                     }
                                 i = j;
@@ -189,7 +182,7 @@ namespace fwdpy11
                       const typename std::vector<T>::const_iterator>
             set_range(const std::vector<T>& v)
             {
-                return { v.cbegin(), v.cend() };
+                return {v.cbegin(), v.cend()};
             }
 
             template <typename T>
@@ -226,14 +219,11 @@ namespace fwdpy11
                     }
                 gsl_matrix_const_view v = gsl_matrix_const_view_array(
                     migmatrix->M.data(), migmatrix->npops, migmatrix->npops);
-                gsl_vector_const_view diag
-                    = gsl_matrix_const_diagonal(&v.matrix);
+                gsl_vector_const_view diag = gsl_matrix_const_diagonal(&v.matrix);
                 bool allequal = true;
-                for (std::size_t i = 0;
-                     allequal == true && i < migmatrix->npops; ++i)
+                for (std::size_t i = 0; allequal == true && i < migmatrix->npops; ++i)
                     {
-                        gsl_vector_const_view rv
-                            = gsl_matrix_const_row(&v.matrix, i);
+                        gsl_vector_const_view rv = gsl_matrix_const_row(&v.matrix, i);
                         double rsum = 0.0;
                         for (std::size_t j = 0; j < migmatrix->npops; ++j)
                             {
@@ -269,9 +259,9 @@ namespace fwdpy11
                             {
                                 if (migmatrix->scaled == true)
                                     {
-                                        auto sum = std::accumulate(
-                                            begin(event.migrates),
-                                            end(event.migrates), 0.);
+                                        auto sum
+                                            = std::accumulate(begin(event.migrates),
+                                                              end(event.migrates), 0.);
                                         if (sum != 0.0 && sum != 1.)
                                             {
                                                 throw std::invalid_argument(
@@ -284,8 +274,7 @@ namespace fwdpy11
                         else if (event.migrates.size()
                                  != migmatrix->npops * migmatrix->npops)
                             {
-                                throw std::invalid_argument(
-                                    "invalid matrix size");
+                                throw std::invalid_argument("invalid matrix size");
                             }
                     }
             }
@@ -313,8 +302,7 @@ namespace fwdpy11
             selfing_rate_change_range selfing_rate_change_tracker;
             migration_rate_change_range migration_rate_change_tracker;
 
-            DiscreteDemography(mass_migration_vector mmig,
-                               set_growth_rates_vector sg,
+            DiscreteDemography(mass_migration_vector mmig, set_growth_rates_vector sg,
                                set_deme_sizes_vector size_changes,
                                set_selfing_rates_vector ssr,
                                std::unique_ptr<MigrationMatrix> m,
@@ -344,19 +332,14 @@ namespace fwdpy11
             // NOTE: needs test.
             {
                 reset_range(mass_migration_tracker, mass_migrations);
-                update_event_times(current_pop_generation,
-                                   mass_migration_tracker);
+                update_event_times(current_pop_generation, mass_migration_tracker);
                 reset_range(growth_rate_change_tracker, set_growth_rates);
-                update_event_times(current_pop_generation,
-                                   growth_rate_change_tracker);
+                update_event_times(current_pop_generation, growth_rate_change_tracker);
                 reset_range(deme_size_change_tracker, set_deme_sizes);
-                update_event_times(current_pop_generation,
-                                   deme_size_change_tracker);
+                update_event_times(current_pop_generation, deme_size_change_tracker);
                 reset_range(selfing_rate_change_tracker, set_selfing_rates);
-                update_event_times(current_pop_generation,
-                                   selfing_rate_change_tracker);
-                reset_range(migration_rate_change_tracker,
-                            set_migration_rates);
+                update_event_times(current_pop_generation, selfing_rate_change_tracker);
+                reset_range(migration_rate_change_tracker, set_migration_rates);
                 update_event_times(current_pop_generation,
                                    migration_rate_change_tracker);
             }
@@ -367,6 +350,18 @@ namespace fwdpy11
             {
                 demographic_model_state_pointer rv(std::move(model_state));
                 return rv;
+            }
+
+            demographic_model_state_pointer&
+            get_model_state_ref() 
+            {
+                return model_state;
+            }
+
+            const demographic_model_state_pointer&
+            get_model_state_ref() const
+            {
+                return model_state;
             }
 
             void